+++ bapccanada/home/admin.py
@@ -1,8 +1,8 @@
 from django.contrib import admin
-from home.models import GPU, Review, Build, UserProfile
+from home.models import GPU, Review, Build
 
 # Register your models here.
 admin.site.register(GPU)
 admin.site.register(Review)
 admin.site.register(Build)
-admin.site.register(UserProfile)
+
